backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.3
pydantic==2.5.0

# Google Sheets Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# Optional PDF Generation (install as needed)
# weasyprint==60.2  # For PDF reports
# playwright==1.40.0  # For JavaScript-enabled PDF reports

# Updated requirements for Pydantic v2
pydantic-settings==2.1.0